Trait magical::blockchain::Blockchain

pub trait Blockchain {
    fn is_online(&self) -> bool;

    fn offline() -> Self;
}

Base trait for a blockchain backend

This trait is always required, even for "air-gapped" backends that don't actually make any external call. Clients that have the ability to make external calls must also implement OnlineBlockchain.

Required methods

fn is_online(&self) -> bool

Return whether or not the client has the ability to fullfill requests

This should always be false for offline-only types, and can be true for types that also implement OnlineBlockchain, if they have the ability to fullfill requests.

fn offline() -> Self

Create a new instance of the client that is offline-only

For types that also implement OnlineBlockchain, this means creating an instance that returns Error::OfflineClient if any of the "online" methods are called.

This is generally implemented by wrapping the client in an Option that has Option::None value when created with this method, and is Option::Some if properly instantiated.

Loading content...